Shader Abstraction finish
